fix: move some files around
