+++ src/app/screens/Onboard/Prompt/index.tsx
@@ -64,11 +64,11 @@ export default function Onboard() {
           <div className="dark:text-white pt-6">
             <div className="mb-2 flex items-center">
               <TwoKeysIcon className="w-7 h-7 mr-2" />
-              <p className="">{t("request1")}</p>
+              <p>{t("request1")}</p>
             </div>
             <div className="mb-2 flex items-center">
               <InfoCircleIcon className="w-7 h-7 mr-2" />
-              <p className="">{t("request2")}</p>
+              <p>{t("request2")}</p>
             </div>
             <div className="mb-2 flex items-center">
               <ClockIcon className="w-7 h-7 mr-2" />
