+++ src/extension/background-script/actions/onboard/index.ts
@@ -1,0 +1,3 @@
+import prompt from "./prompt";
+
+export { prompt };
