+++ src/extension/background-script/router.ts
@@ -7,7 +7,7 @@ import * as ln from "./actions/ln";
 import lnurl, { auth } from "./actions/lnurl";
 import * as mnemonic from "./actions/mnemonic";
 import * as nostr from "./actions/nostr";
-import onboard from "./actions/onboard/onboard";
+import * as onboard from "./actions/onboard";
 import * as payments from "./actions/payments";
 import * as permissions from "./actions/permissions";
 import * as settings from "./actions/settings";
@@ -84,7 +84,7 @@ const routes = {
   // Public calls that are accessible from the inpage script (through the content script)
   public: {
     webbtc: {
-      onboard: onboard,
+      onboard: onboard.prompt,
       enable: allowances.enable,
       getInfo: webbtc.getInfo,
       getAddressOrPrompt: webbtc.getAddressOrPrompt,
@@ -105,13 +105,13 @@ const routes = {
       request: ln.request,
     },
     liquid: {
-      onboard: onboard,
+      onboard: onboard.prompt,
       enable: allowances.enable,
       getAddressOrPrompt: liquid.getAddressOrPrompt,
       signPsetWithPrompt: liquid.signPsetWithPrompt,
     },
     nostr: {
-      onboard: onboard,
+      onboard: onboard.prompt,
       enable: allowances.enable,
       getPublicKeyOrPrompt: nostr.getPublicKeyOrPrompt,
       signEventOrPrompt: nostr.signEventOrPrompt,
